Fix NDVI extraction for L8 bands
This is a regression introduced in 1.5. The old working code truncated
the band values to int, while the 1.5 version tried to do the
computations on float. This should have been better, but non-integer
values come from resampled L8 bands and the no data comparisons no
longer worked.

Besides this, clamp negative reflectance values to 0.