Fix supervised crop mask crash on tiles between strata
Fix a regression from when red-edge band support was implemented in the
crop mask processor. The issue is that GetOutput() will be called
multiple times (with a different set of output dates) and OTB does not
keep strong references to the parent filters in the graph.

This made the processor crash because of an empty input region on tiles
spanning multiple strata.